refactor: findHint method move to hint class
